Final updates plus adding a main file

+++ ft_printf_others.c
@@ -6,15 +6,15 @@
 /*   By: mmendiol <mmendiol@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/27 16:19:18 by mmendiol          #+#    #+#             */
-/*   Updated: 2023/11/27 19:45:37 by mmendiol         ###   ########.fr       */
+/*   Updated: 2023/11/27 16:27:09 by mmendiol          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_putnbr_hexa(unsigned int nb, char *base)
+int	ft_putnbr_hexa(unsigned long nb, char *base)
 {
-	int counter;
+	int	counter;
 
 	counter = 0;
 	if (nb >= 16)
@@ -27,10 +27,10 @@ int	ft_putnbr_hexa(unsigned int nb, char *base)
 	return (counter);
 }
 
-int	ft_putptr(unsigned int nb)
+int	ft_putptr(unsigned long nb)
 {
 	char	*base;
-	int counter;
+	int		counter;
 
 	base = "0123456789abcdef";
 	counter = 0;
